Change.
1. Fixed some logic errors
2. Fixed some display exceptions
3. Redundant code removed
4. Annotations have been added
5. Added file import and export function (.paint .png .jpeg .tiff)
